watchman: improve handling server spawn failure on win32
Summary:
There are some situations where the watchman CLI
would appear to hang forever (or 2 minutes, which is long
enough to consider as forever).  These correlate with
problems spawning the server.

This commit improves this situation by waiting 500ms after
launching the server to see if it has terminated.  If so
then we'll exit with a reasonably clear, but not perfect,
error message.

The wait is acceptable in this code path: we should only
get here if we cannot reach an already running server,
so we are definitely already in a slow path, and 500ms
seems like a reasonably effective grace period for this.

Reviewed By: genevievehelsel

Differential Revision: D22610516

fbshipit-source-id: eea235b82270cd9293e2670eb02d8e494b499d48