watchman: cache Mercurial::getFilesChangedSinceMergeBaseWith results
Summary:
This commit uses the same technique as we use for the mergeBase
cache in the prior commit to produce a cache key, so that we'll effectively
invalidate the results when the dirstate is changed.

This provides both caching and thundering herd protection when multiple
clients are performing SCM aware queries concurrently.  We've had some
issues in the past where we'd OOM a cgroup slice under the combined
memory pressure of the the concurrent `hg` processes.

Reviewed By: chadaustin

Differential Revision: D22582273

fbshipit-source-id: 19f6a81b1c18831878f4a7aab4ec4a1ef82033da