watchman: add default fetch timeout for LRUCache
Summary:
in T34529442 we saw a situation where we were blocked forever
waiting on the result of a cache lookup, and we resolved to add some
kind of timeout.

This commit introduces a timeout that applies to every LRUCache
lookup operation.  It defaults to 5 minutes in duration which should
be long enough as to be equivalent to "forever" in terms of the
amount of time that a client might want to wait for results in
extreme situations that it seems safe to make this a blanket
default.  This applies to the content hash and symlink
target caches.

For the SCM/Mercurial.h operations that use the newer config
based constructor, we have a default of 1 minute for the associated
source control operations.

Reviewed By: genevievehelsel

Differential Revision: D22590189

fbshipit-source-id: 82b0aff76025eee35f9080236ede27273128c545