test for UDS support on Windows instead
Summary:
Windows does not give correct version information when the application is not declared as compatible with Windows 10 in the manifest file. This has cause us to incorrectly disable UNIX domain socket support on Windows 10. (see https://docs.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-verifyversioninfoa#remarks)

Instead we directly test if we can create a UNIX domain socket on Windows and disable it according to the test result.

Reviewed By: wez

Differential Revision: D23772443

fbshipit-source-id: f3895d968a4a0382809cfd358459f7448bae7575