watchman: improve checks for an edenfs mount point
Summary:
We've seen occasional instances where watchman has decided
to use the inotify watcher on an eden mount.  This is undesirable because
the inotify watcher performs an `O(repo)` crawl, which is one of the
pathological worst cases for EdenFS.

It's not clear how watchman gets into that state; one theory is that
watchman is started before EdenFS, but at FB we only allow watching
the filesystem when the `.watchmanconfig` file is present, and that
isn't there if the EdenFS mount is not yet online.

So far the EdenFS watcher is triggered only when we detect something
that looks sufficiently like an EdenFS mount.  This commit adds some
logic to check whether it looks like an offline EdenFS mount.

On POSIX systems, EdenFS will symlink a readme file into the mount
point so that an offline mount doesn't appear totally empty.

This commit looks for that file in the watched root; if it is present
then we assume that this particular location should only be watched
using the EdenFS watcher and then generate a terminal error to prevent
using any of the fallback watchers.

Reviewed By: genevievehelsel

Differential Revision: D23829169

fbshipit-source-id: 2b91a00707c1b670929914b758d2bb7d52a4c136