watchman: Improve error message for caught exceptions
Summary:
This bit of code catches exceptions thrown while
dispatching commands in the watchman server.

Previously it would log just the `std::exception::what` field,
which made it difficult to run down the source of an error
thrown by what I now know is the `fmt` library.

This commit changes this logic to use a helper function from
folly that includes the exception class name in the returned
string.

Reviewed By: genevievehelsel

Differential Revision: D24028202

fbshipit-source-id: 24413d26c568180f787d891f4d1eca1007160f7e