watchman: change suffix queries to be implemented in terms of glob
Summary:
suffix queries are a bit of a niche function; there is only a single
use case inside FB after several years of availability, and can in most
cases be rephrased in terms of a recursive glob.

suffix queries are potentially faster, with `O(number-of-files-with-suffix)`
vs. `O(number-of-files-in-repo)` number of candidates to match against,
provided that a prior `O(number-of-files-in-repo)` walk has completed
prior to the query.  The greater the difference between these two numbers,
the greater the performance difference.

With EdenFS we don't build a suffix index as the act of building it
will cause data for the whole repo to be eagerly fetched for every
watchman startup.

I'd like to remove the suffix index from watchman for a couple of
reasons:

* Improved parity between the EdenFS and non-EdenFS watchers
* Saves `O(number-of-files-ever-observed * 2 * sizeof(void*))` memory
  that is "never" used
* Simplifies future work to improve concurrent read/write access
  to the InMemoryView datastructures by avoiding building a read/write
  suffix index. (laziness!)

This commit causes the query parser to re-phrase suffix queries in
terms of glob queries.

A global configuration option can be used to restore the prior
behavior in case we quickly need to restore that behavior, but
I'm not anticipating that this will be a problem in practice.

Reviewed By: genevievehelsel

Differential Revision: D24028274

fbshipit-source-id: 63324e6056326c5bf65374946828f1b898414e74