watchman: fixup path generator test case
Summary:
For reasons that aren't super clear to me, this test failure
bisected to D23858659 (https://github.com/facebook/watchman/commit/939250d81a4ee89da9cc515dc8182852a39e8be4).  The issue is that the query results include
the `.eden/client` symlink and the logic in this test assumed that
there was only a single symlink in the repo.

Fixup the test to filter out the `.eden` directory from the results.

Reviewed By: genevievehelsel

Differential Revision: D25320978

fbshipit-source-id: 57b00adaecc1e9b33c65f3f31900b337de8a3e18