win: chdir to the filesytem root before daemonizing
Summary:
On Windows, the current working directory of a process cannot be removed by
other processes. When watchman is invoked by Buck or other tools, this leads to
watchman being spawned in the same directory it was invoked from, thus holding
a handle to that directory.

When using EdenFS, clones are cheap and can be removed and recreated fairly
quickly, but with watchman holding a handle to one of its subdirectories, this
causes the removal of the clone to always fail.

The solution is simple, we just need to spawn watchman outside of the
repository.

Reviewed By: fanzeyi

Differential Revision: D25832564

fbshipit-source-id: 98ac9e04693fa797613fffc6c83fd9fb3965dc0c