Stop holding a lock when running a subscription callback
Summary:
We don't need to hold a lock during a subscription callback, and we have no idea what each subscription callback does. So let's not hold the lock while running the callbacks.

I'd also be happy to change this mutex to a `folly::Synchronized` and put protected data inside it, but that's a more elaborate change that I'm unsure people actually want.

Reviewed By: chadaustin

Differential Revision: D26011091

fbshipit-source-id: 8c012317bc73d64f46c9fd8170fcbecd2a0ffad6