add omit_changed_files option to queries
Summary: This diff adds a `omit_changed_files` option to queries, which tells Watchman to not bother calculating and returning the list of changed files from a query. This is particularly useful if a consumer wishes to find the saved-state, but does not care about the files changed, or the client wants to generate this list themselves.

Reviewed By: chadaustin

Differential Revision: D26134034

fbshipit-source-id: 019bc4e9eab3a390b8b3134c25cef407e65e585b