allow opting out of deelevation on startup
Summary:
Some windows sandcastle jobs were failing because we were deelevating.
The jobs are running watchman elevated, and deelevating looses watchman
permission to read the repos it is suppose to be watching which are also
created elevated.

Since sandcastle has a tendency to run things as root and permissions are
complicated there, let's not deelevate there. Deelevating is intended for the
case when a user accidentally runs watchman as root anyways, i.e. intended
for corp. We are defaulting to not deelevating since configuration is also
complicated on sandcastle, some jobs setup there own configuration files, so
it will be safer to leave this configuration off and only enable it on corp (D26240911).

Reviewed By: fanzeyi

Differential Revision: D26104891

fbshipit-source-id: 045939e0ea5ceff289c4384276c98b6a559ea32a