retry daemon restart if an existing daemon is shutting down
Summary:
`watchman shutdown-server; watchman get-sockname` would faily reliably
time out, because the second process would attempt to spawn a new
daemon, but fail to acquire the pidfile lock because the old process
was still shutting down.

One option is to make shutdown-server synchronous, but that doesn't
solve the scenario where commands are running in parallel with
shutdown-server.

Instead, acquire the pidfile lock before fork(), and allow retrying
the spawn attempt if that fails.

Reviewed By: kmancini

Differential Revision: D26611540

fbshipit-source-id: c912ab80c01e5d9085f11802010a6fb3a5280dad