fsevents: enqueue events onto a vector<vector<>>
Summary:
We've seen a lot of cases recently where watchman is forced to recrawl the
entire repository due to receiving kFSEventStreamEventFlagUserDropped events.
These can be due to either fseventsd not being fast enough, or to watchman not
being fast enough to service the callbacks.

Looking at the code, enqueing the events is a O(n) operation due to moving all
the received events onto the lock protected queue of events. We could speed
this up by simply storing events in a vector<vector<>> as a push_back operation
is a constant one, thus speeding up servicing events.

Reviewed By: chadaustin

Differential Revision: D26887863

fbshipit-source-id: 73636b77a559981863968b6fa550c5b381518ad4