fsevents: do not request per-file notification
Summary:
On macOS, bursty applications that writes a large number of files very quickly
can lead to overwhelming fsevents causing it to drop events. While at this
point it is not clear if fsevents or watchman are too slow to service the
events, the result is the same: watchman needs to recrawl the watched
directory. Due to the size of our repository, this can take several minutes.
This behavior was seen while running `hg update` and while compiling with Buck.
In the case of Buck, a bug in the CMake build system caused us to never exclude
the buck-out directory, causing extra unecessary notifications.

To alleviate these issues, the fsevents watcher is modified to only request
directory level changes, and not file level changes when fsevents_watch_files
is set to false. Apple documentation states about the file level changes:

> Request file-level notifications. Your stream will receive events about
> individual files in the hierarchy you're watching instead of only receiving
> directory level notifications. Use this flag with care as it will generate
> significantly more events than without it.

Which appears to match what we've been seeing.

Removing the per-file notification also implies that watchman need to be taught
to always rescan the directory entry when a notification for a directory is
sent and thus this change trades less notification for more `stat(2)` calls.

Several alternative solutions were looked at:
 1) Using kqueue exclusively instead of FSEvents
 2) Moving the buck-out directory outside of the repository
 3) Using a watcher that uses a mix of kqueue+fsevents

All of these have their own flaws and issues. The first one simply crashes
macOS due to watchman exhausting the number of filedescriptors on the system,
raising the limit (from 1M) didn't appear to help. Moving buck-out would help
with Buck writing lots of data, but it would also break a large amount of build
target that assume buck-out is in the repository, and while this is an effort
worth pursuing it will take a while to achieve. Moving buck-out would also not
help with `hg update`. Lastly the last solution on top of being complex and not
fitting well with the rest of the watchman code does have cookie
synchronisation issues.

Reviewed By: chadaustin

Differential Revision: D27083892

fbshipit-source-id: 912d0e2f5ffec128ae68ef5b1b7e432464e2fed2