CookieSync: enable the use of multiple cookies
Summary:
Up to know, a watchman watch covered an entire directory hierarchy, and events
were assumed to be serialy received. Thus, to avoid races in between a watchman
query, and writes to the directory hierarchy, a single cookie was written to
the directory and once observed, it was assumed that all the events that
happened prior to the query had been observed by watchman.

Unfortunately, this no longer holds true if the directory hierarchy is watched
by multiple watches. For instance, if the root items are being watched with
kqueue, but the directories are watched by different FSevent watches, writing a
single cookie to the root will not synchronize the directories. For this, a
cookie needs to be written to the directories and wait until these are
observed.

Reviewed By: genevievehelsel

Differential Revision: D27541885

fbshipit-source-id: 54f0feea7282d23dac64868c16a615ab840db2b8