watcher: move watch cancellation from watcher to notifythread
Summary:
In an upcoming change, the fsevents watch may be used as a nested one instead
of on the whole directory hierarchy. In this case, we want to make sure that
when the directory that is being watched is removed the whole watch isn't
invalidated, only itself. Due to this, let's move the cancellation one level
up.

Reviewed By: fanzeyi

Differential Revision: D27544318

fbshipit-source-id: 06ce1422d4752fe3251c0afc1d51768330666ff2