watcher: add a kqueue+fsevents watcher
Summary:
On macOS, bursty applications that writes a large number of files very quickly
can lead to overwhelming fsevents causing it to drop events. While at this
point it is not clear if fsevents or watchman are too slow to service the
events, the result is the same: watchman needs to recrawl the watched
directory. Due to the size of our repository, this can take several minutes.
This behavior was seen while running `hg update` and while compiling with Buck.
In the case of Buck, the buck-out directory should have been ignored and thus
it shouldn't have affected watchman, but alas, the ignore list doesn't prevent
heavy writes to buck-out to not trigger a dropped event (which is likely a bug
in Apple's software).

To alleviate these issues, a new watcher is introduced that makes use of kqueue
for the top-level files and the root directory, and fsevents for the top-level
directories. Having one fsevent per top-level directory allows for 2 benefits:
 1) Dropping event in one directory doesn't force watchman to recrawl the entire
    repository
 2) We can avoid adding a watcher on the buck-out directory, which would
    prevent fsevents from sending a dropped event.

As an alternative solution, using kqueue exclusively instead of fsevents was
explored, with the unfortunate result of killing macOS, as it pushed macOS
above the maximum number of file descriptors in the system. Another solution
that is being explored is to move the build directory outside of the
repository, but this effort is likely to take months as several build targets
are making the assumption that the build directory is contained in the
repository, it would also not help with `hg update`.

Reviewed By: fanzeyi

Differential Revision: D27030761

fbshipit-source-id: f945a39c8b1c74a59cc9efec475167f60d957d80