recrawl: add an alternative way of recrawling
Summary:
The current way of recrawling only works for non-split watchers as they recrawl
the entire repository, and not a specific given hierarchy. For the
kqueue+fsevents recrawl, the flag W_PENDING_RECURSIVE was used to force a
recrawl of the watched directory, but this was actually racy with cookies being
created on disk as the recursive crawler would always consider cookies and thus
cookies could be observed before the rest of the hierarchy was fully crawled.

To solve this, a watcher can indicate that it is desynced with
W_PENDING_IS_DESYNCED and the crawler will both not consider cookies when set,
and will abort all cookies after processing all of the desynced paths. This is
in line with what the current recrawl logic does.

In the future, we can replace the old recrawl logic with this one for all the
watcher, simplifying the code a bit.

Reviewed By: wez

Differential Revision: D27767604

fbshipit-source-id: f6bb9ff56ea344952e444e29f7624403713b8989