Bad pending changes condition
Summary:
Pending Changes should be removed from both the list and the tree at the same time.
From auditing the PendingChanges code for when things are removed, it looks like this
is true. Additionally, this case is not hit in our integration tests.

If changes are removed from the list but not the tree, we could see bad things like
missing changes. This should be removed.

Reviewed By: chadaustin

Differential Revision: D28132160

fbshipit-source-id: e23dc6a70e6a5f845c88880a0eee05c104a3aa26