watchers: fix condition variable misuse
Summary:
Condition variables are terrible synchronization mechanism that are very easy
to misuse. In these cases, another thread may have modified the shared variable
while no waiters were waiting on the condition variable. What happens then is
that once the condition variable is waited on, it will simply block, regardless
of whether the condition was previous notified.

The fix is trivial: we just need to check the condition expression before
waiting on the condition variable.

Reviewed By: chadaustin

Differential Revision: D29406650

fbshipit-source-id: 29768b71b5f19e3b7616dd9281ae73ab425b37ea