fix a Watchman divergence in FSEvents directory-only mode
Summary:
I don't have enough logging and tracing to precisely determine the
sequence of events that leads to a divergent state, but I have a
reproduction that occasionally results in a state where all Watchman
queries produce incorrect data henceforth. I can reproduce this in the
dirfsevents watcher (when fsevents_watch_files is set to false in
/etc/watchman.json).

My hypothesis is that, when a crawl and a pending directory change
event overlap, the crawl may coalence the pending directory change
incorrectly, resulting in Watchman not calling stat() on each child of
the directory.

Instead of a per-watcher flag that indicates whether the watcher
supports per-file or per-directory change events, instead define a
three-state flag hierarchy on the pending event (non-recursive,
shallow, and deep) and coalesce it properly in PendingCollection.

This may increase the number of stats that Watchman does, and perhaps
cause queries to run more slowly, but the old logic was incorrect.

Reviewed By: xavierd

Differential Revision: D29412921

fbshipit-source-id: 1fc524de383bdae70661ce58883d7f38414d0aa9