watchman_client: remove "generic" error
Summary:
I'd like to be able to introspect the errors returned by Watchman client, and
in particular I'd like to know which errors are the result of losing connection
to Watchman.

I'd also like to be able to differentiate between errors that caused the worker
task to crash, and errors that are returned to the caller, so this splits the error
enum accordingly.

Let's report those accordingly.

Reviewed By: kassens

Differential Revision: D29550641

fbshipit-source-id: 1ffcf6d676ca1175fd40bf42aeb6f770ed6ca544