watchman/serde_bser: update to anyhow / thiserror
Summary:
This was using error_chain, which is pretty annoying to use in an async context
because its errors aren't Sync. This updates it to use anyhow and thiserror,
which are a) more common

The existing code was doing a bit of mixing of strongly-typed errors and actual
typed errors. I updated the internal-facing bits (reader) that want to use
strings for their errors and add context to them to use anyhow, and the
external-facing bits to use thiserror.

Note that the error type isn't actually public in this crate so it cannot be
matched on by callers even though it's now actually an enum.

Reviewed By: chadaustin

Differential Revision: D29555272

fbshipit-source-id: 4465ac8d436ac5fb60cc9d2cb73fdcfa65e73339