watchman/serde_bser: fix rust-2018-idioms lints
Summary:
Like it says in the title. This fixes 2 lints:

- https://doc.rust-lang.org/rustc/lints/listing/allowed-by-default.html#elided-lifetimes-in-paths
- https://doc.rust-lang.org/rustc/lints/listing/allowed-by-default.html#explicit-outlives-requirements

Weirdly, Rust complains about this:

```
warning: outlives requirements can be inferred
   --> src/de/read.rs:223:30
    |
223 |   pub enum Reference<'b, 'c, T>
    |  ______________________________^
224 | | where
225 | |     T: ?Sized,
    | |______________^ help: remove these bounds
```

There is no lifetime requirements here, but switching this to `T: ?Sized` fixes
it, hence the seemingly-opinionated-but-actually-required rewriting of this
declaration here.

This is useful because we have some code here (see: next diff in this stack)
that disallows those via a Cargo config.

Reviewed By: kassens

Differential Revision: D29611430

fbshipit-source-id: 2eb0126e680c01387cdb6ef69e5f061c2990d058