watchman/rust: properly handle subscription cancellation
Summary:
The current code doesn't work because we can't successfully parse a
subscription cancellation at the point where we try to do it because we also
expect some other fields to be present that aren't there on this notification.

Instead, we should peek at the canceled field at the same time we peek for
unilateral PDUs. This diff does that.

See here for a bug report:

https://fb.workplace.com/groups/watchman.users/permalink/2910791889238279/

Reviewed By: kassens

Differential Revision: D29636575

fbshipit-source-id: 8804fac50c1de46d5873e396790501bc439bbc8f