fix for uncommitted git changes
Summary:
Git::getFilesChangedSinceMergeBaseWith was doing `git diff --name-only commit...`.

```
This form is to view the changes on the branch containing and up to the second <commit>, starting at a common ancestor of both
<commit>.  git diff A...B is equivalent to git diff $(git merge-base A B) B. You can omit any one of <commit>, which has the same
effect as using HEAD instead.
```

since `commit` is already the mergebase, we don't need the first part of this behavior.

but more importantly, since the 2nd commit is omitted, it uses `HEAD`, which only shows _committed_ changes between the mergebase.

if we leave off the `...` we compare against the working copy:

```
This form is to view the changes you have in your working tree relative to the named <commit>. You can use HEAD to compare it with
the latest commit, or a branch name to compare with the tip of a different branch.
```

Reviewed By: chadaustin

Differential Revision: D30513337

fbshipit-source-id: 03012e895fd695c378696d1ef8cb43bff26d5a34