enable EdenFS support by default in open source build
Summary:
The open source builds have been failing for a while, because
test_capabilities assumes the EdenFS watcher is available.

The ENABLE_EDEN_SUPPORT option was originally added to reduce build
times in the open source build, but we already unconditionally build
fbthrift and fb303. To see if this unblocks the releases, default this
option on. Eventually, we should remove it, and if open source builds
are a problem, we should address that directly, perhaps by adding a
non-Thrift protocol to communicate with EdenFS.

Reviewed By: fanzeyi

Differential Revision: D31635078

fbshipit-source-id: 4369e7a7270a7fdffc8d6df009203d553be6df69