eden: include files if they are specified in `path` parameter
Summary:
With EdenFS watcher, if the user sends a query with `path` parameter and the path itself is a file, we won't include it when we return the result. This behavior does not match what we have with regular filesystem watchers. The root cause is that we directly let EdenFS to resolve the glob `<path>/**/*` then process the result. This glob query won't match the path itself if it is a file.

In this diff, we fix this mismatch by adding a separate glob of the path itself when we ask EdenFS to resolve them. Then we make sure we skip processing directories after EdenFS gives us result. This should take care of this case.

Reviewed By: chadaustin

Differential Revision: D31633962

fbshipit-source-id: c66cf9eaf3fa39bd7075b66b666eb18f22f2fbb0