unblock queries after the directory has been scanned
Summary:
NB: Intentionally, this is an incomplete fix. But it's low-risk and fixes
the first unit test in this area. The bigger fix will come next.

If an event arrives from a watcher that doesn't have per-file watching
in the same directory, we need to wait until we walk the entire
directory before we unblock the query.

Reviewed By: kmancini

Differential Revision: D32188270

fbshipit-source-id: 39551298b210c40d4a69e9f48a4ff6c5feb057e1