consider a ping without a new item as remaining settled
Summary:
The iothread can be woken by items being added from the notifythread
or by a ping. There's no need to distinguish between these, so remove
the ping flag and only check the number of items pulled from the
queue.

This is important for the next diff, because pings that don't add
items should not be considered unsettling.

Reviewed By: kmancini

Differential Revision: D32748874

fbshipit-source-id: 5e43855d58704381e452c28963efdd097b71b618