fix test_notify_dir on macOS by replacing propagateToParentDirIfAppropriate with synthesized event
Summary:
This logic makes it hard to reason about the behavior of the IO
thread. It looks like it tries to be precise but instead let's just be
overly conservative by synthesizing parent directory event changes in
the watchers that don't provide them.

This change is behavior-preserving because it was already broken on
macOS. The next diff will fix macOS.

Reviewed By: xavierd

Differential Revision: D34486515

fbshipit-source-id: 027dc979f69b65a9050c77f969770eaebfb0db4e