docs: add social banner in support of Ukraine (#1006)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

Pull Request resolved: https://github.com/facebook/watchman/pull/1006

Reviewed By: fanzeyi

Differential Revision: D34648013

Pulled By: dmitryvinn-fb

fbshipit-source-id: 48b78070fbf0eb2332132c9b2b56fddae30991f0