skip the fundamentally flaky test_watchReap
Summary:
test_watchReap was quite flaky on my devserver, and after some
analysis, it looks like it's fundamentally flaky, as a three-second
reap time is not sufficient under heavy contention, and a longer reap
time would make the test that much slower.

This logic would be better expressed as a C++ test that can fake the
clock.

We don't use automatic watch reaping anyway, so simply disable the
test for now.

Reviewed By: genevievehelsel

Differential Revision: D34655479

fbshipit-source-id: 2ea26d4abb84e9ca17f451615e5e0021805afdd1