fix some flaky integration tests
Summary:
Tests that start watchman, expect it to fail, and read its output want
to be able to see the entire output before the process exits. Some
tests were causing subprocess.kill to be called before the output was
entirely written to the CLI log. To fix those flaky failures, wait for
watchman to fail cleanly rather than sending SIGKILL.

Reviewed By: genevievehelsel

Differential Revision: D34655170

fbshipit-source-id: bffb0b1a89169f9df0f3e9ebfa6ca3b986fed978