eden: stop subscription from fetching the entire repo
Summary:
On a large monorepository, daily checkout operation can jump over thousands of
commits, that modify a very large amout of files. Unfortunately, the design of
since queries and subscription is that they need to return all the files that
have been modified in this checkout operation in order for tools to potentially
invalidate their caches.

When on EdenFS, this unfortunately means that Watchman will force EdenFS to
fetch data about all of these files post-update, which also include files that
aren't in the working set.

Well behaving tools already know how to deal with fresh instance responses and
are requesting them to be empty, thus if we detect that the set of files is too
big, we can simply force Watchman to return a fresh instance instead of
fetching data for a very large amount of files.

Reviewed By: chadaustin, jtbraun

Differential Revision: D35322477

fbshipit-source-id: a147a0ef835612bdb9b0564da0ff7f762c63092a