introduce a TypedCommand interface for registering commands
Summary:
TypedCommand will allow command definitions to specify their interface
in the form a Request and Response struct, each of which can be
serialized to and from JSON (or BSER). This will make it easier for
CommandDefinition to have some client-side logic for rendering the
response in a human-readable format.

Reviewed By: fanzeyi

Differential Revision: D35298220

fbshipit-source-id: 9eddf5b71e10d5759806fd4bd9f15c0cd87fa29b